Move purely PSI checks of contract presence to ktPsiUtil.kt
